+++ src/pages/dashboard/[dashboardId]/edit.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import DashboardEdit from '@/containers/Dashboard/Edit';
-import Layout from '@/components/Layout';
+import Layout from '@/components/layouts/Layout';
 
 function DashboardEditPage() {
   const router = useRouter();
