+++ src/pages/dashboard/[dashboardId]/index.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import { useState } from 'react';
 import useRequest from '@/hooks/useRequest';
 import Dashboard from '@/containers/Dashboard/DashboardId';
-import Layout from '@/components/Layout';
+import Layout from '@/components/layouts/Layout';
 import { MembersProps } from '../../api/mock';
 
 function DashboardPage() {
