+++ src/pages/test.tsx
@@ -1,0 +1,11 @@
+import Layout from '@/components/layouts/Layout';
+
+function Test() {
+  return (
+    <Layout>
+      <div className='h-screen w-full bg-black'>TEST</div>
+    </Layout>
+  );
+}
+
+export default Test;
